[#43] Refactor database to storage
