+++ lib/di/module/storage_module.dart
@@ -2,7 +2,7 @@ import 'package:injectable/injectable.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 @module
-abstract class DatabaseModule {
+abstract class StorageModule {
   @singleton
   @preResolve
   Future<SharedPreferences> get sharedPreferences =>
