+++ lib/usecases/login_use_case.dart
@@ -1,7 +1,7 @@
 import 'package:injectable/injectable.dart';
 import 'package:lydiaryanfluttersurvey/api/repository/auth_repository.dart';
-import 'package:lydiaryanfluttersurvey/database/shared_preferences_utils.dart';
 import 'package:lydiaryanfluttersurvey/model/response/login_response.dart';
+import 'package:lydiaryanfluttersurvey/storage/shared_preferences_utils.dart';
 import 'package:lydiaryanfluttersurvey/usecases/base/base_use_case.dart';
 
 class LoginInput {
